moved landingPage code to index.
